requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
pydantic==2.7.0
python-dotenv==1.0.1
httpx==0.27.0
pyreadstat==1.2.7
pandas==2.2.2
